All tests pass locally

+++ code/fraction/src/model/fraction.py
@@ -36,8 +36,8 @@ class Fraction:
         if q == 0:
             raise InvalidFractionError('q cannot be equal to zero')
         common_divisor = rational_math.gcd(p, q)
-        self.p = p / common_divisor
-        self.q = q / common_divisor
+        self.p = int(p / common_divisor)
+        self.q = int(q / common_divisor)
 
     def __str__(self):
         return '{p}/{q}'.format(p=self.p, q=self.q)
